remove traces of license icons, rename components and clean up to remove license icon functionality
